# docs/api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Artel Orchestration Server API",
    "description": "API contract for SDK registration and agent action delivery through the Artel orchestration server.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "/",
      "description": "Same origin as the request"
    }
  ],
...
  "paths": {
...
    "/api/projects/{projectId}/game-builds/{gameBuildId}/content-map/scan": {
      "post": {
        "tags": [
          "Content Map"
        ],
        "summary": "원격 근거 스캔 요청",
        "description": "이 빌드를 실행 중인 게임에 근거 스캔을 시킨다. 즉시 202 로 답하며, 완료는 조회 API 의 `lastScan` 으로 확인한다. 붙어 있는 게임이 없으면 409.",
        "operationId": "startScan",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "프로젝트 id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "gameBuildId",
            "in": "path",
            "description": "게임 빌드 id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/StartContentMapScanResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/projects/{projectId}": {
      "get": {
        "tags": [
          "Project"
        ],
        "summary": "프로젝트 상세",
        "description": "참여자가 아니면 404.",
        "operationId": "get_1",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "프로젝트 id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetailResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Project"
        ],
        "summary": "프로젝트 삭제",
        "description": "소유자만 가능하다. 실제로 지우지 않고 삭제 표시만 한다.",
        "operationId": "delete_1",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "프로젝트 id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteProjectResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Project"
        ],
        "summary": "프로젝트 수정",
        "description": "보낸 필드만 반영한다. 설명을 지우려면 빈 문자열을 보낸다.",
        "operationId": "update_1",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "프로젝트 id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetailResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/projects/{projectId}/game-builds/{gameBuildId}/content-map": {
      "get": {
        "tags": [
          "Content Map"
        ],
        "summary": "씬 명세 조회",
        "description": "이 빌드에 앉은 씬·기능·전이·gap 을 한 번에 읽는다. `contentMap` 이 null 이면 등록된 문서가 없고, `contentMap.ingestedAt` 이 null 이면 등록만 되고 아직 앉지 않았다.",
        "operationId": "read_1",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "프로젝트 id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "gameBuildId",
            "in": "path",
            "description": "게임 빌드 id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ContentMapResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{projectId}/game-builds/{gameBuildId}/content-map/events": {
      "get": {
        "tags": [
          "Content Map"
        ],
        "summary": "스캔·적재 진행 SSE",
        "description": "이 빌드의 스캔 상태와 문서 적재 진행을 push 로 흘린다. 구독 직후 `snapshot` 프레임이 한 번 오고, 그 뒤로 상태가 바뀔 때마다 `scan` · `ingest` · `document` 가 온다. 서버가 먼저 끊지 않는다.",
        "operationId": "events_1",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "프로젝트 id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "gameBuildId",
            "in": "path",
            "description": "게임 빌드 id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ServerSentEventContentMapStreamEvent"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "JsonNode": {
        "type": "object"
      },
...
      "DocumentUploaderResponse": {
        "required": [
          "displayName",
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          }
        }
      },
      "ProjectDetailResponse": {
        "required": [
          "createdAt",
          "genre",
          "id",
          "myRole",
          "name",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "genre": {
            "type": "string",
            "enum": [
              "ACTION",
              "RPG",
              "PUZZLE",
              "SIMULATION",
              "STRATEGY",
              "SPORTS",
              "SHOOTER",
              "CASUAL",
              "OTHER"
            ]
          },
          "document": {
            "$ref": "#/components/schemas/ProjectDocumentResponse"
          },
          "myRole": {
            "type": "string",
            "enum": [
              "OWNER",
              "MEMBER"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ProjectDocumentResponse": {
        "required": [
          "contentType",
          "fileName",
          "id",
          "parseStatus",
          "sizeBytes",
          "uploadedAt",
          "uploadedBy",
          "version"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "fileName": {
            "type": "string"
          },
          "contentType": {
            "type": "string"
          },
          "sizeBytes": {
            "type": "integer",
            "format": "int64"
          },
          "uploadedAt": {
            "type": "string",
            "format": "date-time"
          },
          "uploadedBy": {
            "$ref": "#/components/schemas/DocumentUploaderResponse"
          },
          "parseStatus": {
            "type": "string"
          }
        }
      },
...
      "StartContentMapScanResponse": {
        "required": [
          "gameInstanceId",
          "gameInstanceName",
          "requestedAt",
          "state"
        ],
        "type": "object",
        "properties": {
          "gameInstanceId": {
            "type": "integer",
            "description": "명령을 받은 게임 인스턴스",
            "format": "int64"
          },
          "gameInstanceName": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "description": "늘 REQUESTED 다. 조회 API 의 lastScan.state 가 여기서 움직이는 것을 본다",
            "enum": [
              "REQUESTED",
              "SUCCEEDED",
              "FAILED"
            ]
          },
          "requestedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "description": "원격 스캔 요청 결과"
      },
...
      "UpdateProjectRequest": {
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 80,
            "minLength": 1,
            "type": "string"
          },
          "description": {
            "maxLength": 2000,
            "minLength": 0,
            "type": "string"
          },
          "genre": {
            "type": "string",
            "enum": [
              "ACTION",
              "RPG",
              "PUZZLE",
              "SIMULATION",
              "STRATEGY",
              "SPORTS",
              "SHOOTER",
              "CASUAL",
              "OTHER"
            ]
          }
        }
      },
...
      "Always": {
        "required": [
          "kind"
        ],
        "type": "object",
        "description": "조건 없음",
        "allOf": [
          {
            "$ref": "#/components/schemas/ConditionNodeResponse"
          }
        ]
      },
...
      "ContentMapEdgeResponse": {
        "required": [
          "fromSceneId",
          "source",
          "toSceneName"
        ],
        "type": "object",
        "properties": {
          "fromSceneId": {
            "type": "integer",
            "format": "int64"
          },
          "toSceneName": {
            "type": "string"
          },
          "toSceneId": {
            "type": "integer",
            "description": "아직 순회하지 못한 씬이면 null",
            "format": "int64"
          },
          "capabilityId": {
            "type": "integer",
            "description": "자동 전이거나 기능이 지워졌으면 null",
            "format": "int64"
          },
          "capabilitySummary": {
            "type": "string",
            "description": "무엇을 해서 그 씬으로 가는가. 계약 밖의 덤이다"
          },
          "givenText": {
            "type": "string",
            "description": "같은 컨트롤이 조건으로 갈릴 때 둘을 가르는 조건. 계약 밖의 덤이다"
          },
          "given": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Always"
              },
              {
                "$ref": "#/components/schemas/Either"
              },
              {
                "$ref": "#/components/schemas/Every"
              },
              {
                "$ref": "#/components/schemas/Gesture"
              },
              {
                "$ref": "#/components/schemas/Test"
              },
              {
                "$ref": "#/components/schemas/Unknown"
              }
            ]
          },
          "source": {
            "type": "string",
            "description": "static · runtime. runtime 은 정적 분석이 놓친 전이다"
          },
          "verifiedAt": {
            "type": "string",
            "description": "null 이면 아직 못 가본 전이 — 커버리지 구멍이다",
            "format": "date-time"
          }
        },
        "description": "씬 전이 하나"
      },
      "ContentMapResponse": {
        "required": [
          "edges",
          "gaps",
          "pendingDocuments",
          "scenes",
          "screenTransitions",
          "verification"
        ],
        "type": "object",
        "properties": {
          "contentMap": {
            "$ref": "#/components/schemas/ContentMapSummaryResponse"
          },
          "scenes": {
            "type": "array",
            "description": "씬 목록. 이름 오름차순",
            "items": {
              "$ref": "#/components/schemas/ContentMapSceneResponse"
            }
          },
          "edges": {
            "type": "array",
            "description": "씬 전이",
            "items": {
              "$ref": "#/components/schemas/ContentMapEdgeResponse"
            }
          },
          "screenTransitions": {
            "type": "array",
            "description": "화면 전이. QA 런이 관측한 것만 있다",
            "items": {
              "$ref": "#/components/schemas/ContentMapScreenTransitionResponse"
            }
          },
          "gaps": {
            "type": "array",
            "description": "명세가 못 된 사유의 분포. QA 결함이 아니라 개발 우선순위 신호다",
            "items": {
              "$ref": "#/components/schemas/SpecGapCountResponse"
            }
          },
          "verification": {
            "$ref": "#/components/schemas/VerificationResponse"
          },
          "pendingDocuments": {
            "type": "array",
            "description": "등록됐지만 아직 앉지 못한 문서",
            "items": {
              "$ref": "#/components/schemas/PendingDocumentResponse"
            }
          },
          "lastScan": {
            "$ref": "#/components/schemas/LastScanResponse"
          }
        },
        "description": "씬 명세(content map) 조회 결과"
      },
      "ContentMapSceneResponse": {
        "required": [
          "capabilities",
          "capabilityList",
          "id",
          "name",
          "origin",
          "screens",
          "steps",
          "walked"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "walked": {
            "type": "boolean",
            "description": "순회했나. false 면 기능이 비어 있는 것이 정상이다"
          },
          "origin": {
            "type": "string",
            "description": "evidence · observed. observed 는 QA 런이 서 보기만 한 씬이라 기능이 비는 것이 정상이다"
          },
          "capabilities": {
            "$ref": "#/components/schemas/SceneCapabilityCountResponse"
          },
          "steps": {
            "type": "array",
            "description": "이 씬의 조작 단계. not-a-step 은 빠진다",
            "items": {
              "$ref": "#/components/schemas/SceneStepResponse"
            }
          },
          "thumbnail": {
            "$ref": "#/components/schemas/SceneThumbnailResponse"
          },
          "screens": {
            "type": "array",
            "description": "이 씬에서 관측된 화면. QA 런 전에는 비고 그것이 정상이다",
            "items": {
              "$ref": "#/components/schemas/ContentMapScreenResponse"
            }
          },
          "capabilityList": {
            "type": "array",
            "description": "capabilities 가 센 그 행들. 개수의 합이 capabilities.total 과 같다",
            "items": {
              "$ref": "#/components/schemas/SceneCapabilityResponse"
            }
          }
        },
        "description": "씬 하나"
      },
      "ContentMapScreenResponse": {
        "required": [
          "capabilities",
          "discriminator",
          "id",
          "observedCount",
          "sceneId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "sceneId": {
            "type": "integer",
            "description": "이 화면이 속한 씬. 화면 전이가 화면 id 로만 오므로 되짚을 자리가 필요하다",
            "format": "int64"
          },
          "name": {
            "type": "string",
            "description": "표시용. 조인 키가 아니다"
          },
          "discriminator": {
            "$ref": "#/components/schemas/JsonNode"
          },
          "observedCount": {
            "type": "integer",
            "description": "이 화면을 몇 번 지나갔나",
            "format": "int32"
          },
          "firstSeenQaRunId": {
            "type": "integer",
            "description": "이 화면을 처음 본 QA 런",
            "format": "int64"
          },
          "image": {
            "$ref": "#/components/schemas/ScreenImageResponse"
          },
          "capabilities": {
            "type": "array",
            "description": "이 screen 에 묶인 capability. scene 의 capabilityList 로 대신 채우지 않는다",
            "items": {
              "$ref": "#/components/schemas/ScreenCapabilityResponse"
            }
          }
        },
        "description": "씬 안의 화면 하나"
      },
      "ContentMapScreenTransitionResponse": {
        "required": [
          "crossesScene",
          "fromScreenId",
          "id",
          "kind",
          "observedCount",
          "toScreenId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "fromScreenId": {
            "type": "integer",
            "format": "int64"
          },
          "toScreenId": {
            "type": "integer",
            "format": "int64"
          },
          "capabilityId": {
            "type": "integer",
            "description": "null 이면 자동 전이 — TC 가 지시할 수 없다",
            "format": "int64"
          },
          "capabilitySummary": {
            "type": "string",
            "description": "무엇을 해서 넘어갔나"
          },
          "kind": {
            "type": "string",
            "description": "action · state · auto"
          },
          "crossesScene": {
            "type": "boolean",
            "description": "씬 경계를 넘었나. false 면 같은 씬 안의 상태 변화다"
          },
          "observedCount": {
            "type": "integer",
            "description": "이 전이를 몇 번 지나갔나",
            "format": "int32"
          },
          "firstSeenQaRunId": {
            "type": "integer",
            "description": "이 전이를 처음 본 QA 런",
            "format": "int64"
          }
        },
        "description": "화면 전이 하나"
      },
      "ContentMapSummaryResponse": {
        "required": [
          "id",
          "rootedBy"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "rootedBy": {
            "type": "string",
            "description": "이 지도를 세운 경로. evidence 는 근거 문서 등록이, observation 은 QA 런의 관측이 세웠다"
          },
          "capture": {
            "type": "string",
            "description": "editor · editor-play · player 중 하나. 근거 문서가 아직 없으면 null"
          },
          "schemaVersion": {
            "type": "integer",
            "description": "근거 문서의 세대. 근거 문서가 아직 없으면 null",
            "format": "int32"
          },
          "evidenceDigest": {
            "type": "string",
            "description": "구워진 `evidence` 전체의 지문. 값이 달라지면 코드가 바뀐 것이다. 근거 문서가 아직 없으면 null"
          },
          "unity": {
            "type": "string"
          },
          "platform": {
            "type": "string"
          },
          "sdkVersion": {
            "type": "string"
          },
          "ingestedAt": {
            "type": "string",
            "description": "마지막 적재 시각. null 이면 등록만 되고 아직 앉지 않았다",
            "format": "date-time"
          }
        },
        "description": "지도 루트"
      },
      "Either": {
        "required": [
          "kind"
        ],
        "type": "object",
        "description": "parts 중 하나만 성립하면 된다",
        "allOf": [
          {
            "$ref": "#/components/schemas/ConditionNodeResponse"
          }
        ]
      },
...
      "LastScanResponse": {
        "required": [
          "gameInstanceId",
          "gameInstanceName",
          "requestedAt",
          "state"
        ],
        "type": "object",
        "properties": {
          "state": {
            "type": "string",
            "description": "REQUESTED · SUCCEEDED · FAILED",
            "enum": [
              "REQUESTED",
              "SUCCEEDED",
              "FAILED"
            ]
          },
          "gameInstanceId": {
            "type": "integer",
            "description": "명령을 받은 게임 인스턴스",
            "format": "int64"
          },
          "gameInstanceName": {
            "type": "string"
          },
          "requestedAt": {
            "type": "string",
            "format": "date-time"
          },
          "finishedAt": {
            "type": "string",
            "description": "끝난 시각. REQUESTED 이면 null",
            "format": "date-time"
          },
          "ingestedDocuments": {
            "type": "integer",
            "description": "이번 스캔이 앉힌 문서 수. SUCCEEDED 인데 0 이면 올라온 문서가 없었다는 뜻",
            "format": "int32"
          },
          "error": {
            "type": "string",
            "description": "FAILED 일 때 사람에게 보여 줄 사유"
          }
        },
        "description": "마지막 원격 스캔의 상태"
      },
      "PendingDocumentResponse": {
        "required": [
          "documentId",
          "receivedAt"
        ],
        "type": "object",
        "properties": {
          "documentId": {
            "type": "integer",
            "format": "int64"
          },
          "receivedAt": {
            "type": "string",
            "format": "date-time"
          },
          "ingestFailedAt": {
            "type": "string",
            "description": "마지막 적재 실패 시각. null 이면 아직 시도하지 않았다",
            "format": "date-time"
          },
          "ingestError": {
            "type": "string",
            "description": "마지막 실패 사유 한 줄"
          }
        },
        "description": "적재를 기다리는 문서"
      },
      "SceneCapabilityCountResponse": {
        "required": [
          "needsProbe",
          "notAStep",
          "runnable",
          "total",
          "unreachablePrecondition"
        ],
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "description": "네 칸의 합",
            "format": "int64"
          },
          "runnable": {
            "type": "integer",
            "description": "조작이 있고 관측 가능한 효과가 있다. 판정까지 자동",
            "format": "int64"
          },
          "needsProbe": {
            "type": "integer",
            "description": "조작은 있는데 무엇이 달라지는지 `evidence` 가 말하지 않는다",
            "format": "int64"
          },
          "notAStep": {
            "type": "integer",
            "description": "조작이 없다. 단독 명세가 아니라 given/then 의 재료다",
            "format": "int64"
          },
          "unreachablePrecondition": {
            "type": "integer",
            "description": "조건은 아는데 그 상태를 만드는 절차가 `evidence` 에 없다",
            "format": "int64"
          }
        },
        "description": "씬 하나의 기능 상태 분포"
      },
      "SceneCapabilityResponse": {
        "required": [
          "actionability",
          "applicability",
          "id",
          "interaction",
          "observability",
          "origin",
          "scenePresence",
          "status",
          "summary",
          "verification"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "capability.id. steps 의 같은 id 와 같은 행이다",
            "format": "int64"
          },
          "summary": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "description": "runnable · needs-probe · unreachable-precondition · not-a-step"
          },
          "origin": {
            "type": "string",
            "description": "evidence · observed · inferred · human. 어디서 알아냈나"
          },
          "verification": {
            "type": "string",
            "description": "unverified · confirmed · contradicted. 실행으로 확인됐나"
          },
          "scenePresence": {
            "type": "string",
            "description": "placed · persistent-evidenced · persistent-unconfirmed. 이 행이 왜 이 scene 에 있나. persistent-unconfirmed 는 여기서 되는지 아직 아무도 안 봤다"
          },
          "actionability": {
            "type": "string",
            "description": "이 조작을 실제로 할 수 있는가"
          },
          "observability": {
            "type": "string",
            "description": "그 결과를 볼 수 있는가"
          },
          "applicability": {
            "type": "string",
            "description": "이 빌드에 이 규칙이 적용되는가"
          },
          "interaction": {
            "type": "string",
            "description": "click · press · none 등. 프로토콜 메서드가 아니라 의도다"
          }
        },
        "description": "씬 하나의 기능 하나"
      },
      "SceneStepResponse": {
        "required": [
          "id",
          "interaction",
          "status",
          "summary"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "capability.id",
            "format": "int64"
          },
          "summary": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "description": "runnable · needs-probe · unreachable-precondition"
          },
          "interaction": {
            "type": "string",
            "description": "click · press · drag · none 등"
          },
          "inputKey": {
            "type": "string",
            "description": "press 일 때의 키 이름"
          },
          "controlLabel": {
            "type": "string",
            "description": "누를 수 있는 것에 쓰인 글자"
          },
          "controlPath": {
            "type": "string",
            "description": "사람이 읽는 계층 경로"
          },
          "givenText": {
            "type": "string",
            "description": "조건 한 줄. 오늘은 전부 null 이다(ARTEL-447)"
          },
          "given": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Always"
              },
              {
                "$ref": "#/components/schemas/Either"
              },
              {
                "$ref": "#/components/schemas/Every"
              },
              {
                "$ref": "#/components/schemas/Gesture"
              },
              {
                "$ref": "#/components/schemas/Test"
              },
              {
                "$ref": "#/components/schemas/Unknown"
              }
            ]
          }
        },
        "description": "씬 하나의 조작 단계"
      },
      "SceneThumbnailResponse": {
        "required": [
          "state"
        ],
        "type": "object",
        "properties": {
          "state": {
            "type": "string",
            "description": "available · unavailable"
          },
          "url": {
            "type": "string",
            "description": "서명된 단기 주소. state 가 available 일 때만 있다"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "width": {
            "type": "integer",
            "format": "int32"
          },
          "height": {
            "type": "integer",
            "format": "int32"
          },
          "reason": {
            "type": "string",
            "description": "못 만든 이유. SDK 가 준 failureCode 를 그대로 옮긴다"
          }
        },
        "description": "이 씬의 대표 이미지. null 이면 `evidence` 가 캡처를 아예 신고하지 않았다"
      },
...
      "SpecGapCountResponse": {
        "required": [
          "count",
          "reason"
        ],
        "type": "object",
        "properties": {
          "reason": {
            "type": "string"
          },
          "count": {
            "type": "integer",
            "format": "int32"
          }
        },
        "description": "명세가 못 된 사유별 집계"
      },
...
      "VerificationResponse": {
        "required": [
          "total",
          "verified"
        ],
        "type": "object",
        "properties": {
          "verified": {
            "type": "integer",
            "description": "verification 이 unverified 가 아닌 기능 수",
            "format": "int64"
          },
          "total": {
            "type": "integer",
            "description": "`evidence` 출신 기능 수",
            "format": "int64"
          }
        },
        "description": "`evidence` 출신 기능의 실행 확인 비율"
      },
      "ServerSentEventContentMapStreamEvent": {
        "type": "object"
      },
...
      "DeleteProjectResponse": {
        "required": [
          "deleted",
          "projectId"
        ],
        "type": "object",
        "properties": {
          "deleted": {
            "type": "boolean"
          },
          "projectId": {
            "type": "string"
          }
        }
      },
...
    }
  }
}